# More efficient comparison of the same and finite numbers