@agdavydov81 2023-07-04
Hi!
The comparison speedup is great news!
Unfortunately, right now I can't review your pull request, but I promise to back in couple days.
In any way thank you very much!