@plokhotnyuk 2023-07-06
For yet more efficient comparison you can replace
```
            final long sig_n_prime_w1 = Mul64Impl.unsignedMultiplyHigh(sig_x, __CY);
            final long sig_n_prime_w0 = sig_x * __CY;
            // if values are equal
            if (sig_n_prime_w1 == 0 && sig_n_prime_w0 == sig_y) {
                return 0;
            }
            // if positive, return whichever significand abs is smaller
            // (converse if negative)
            return (sig_n_prime_w1 == 0 && UnsignedLong.isLess(sig_n_prime_w0, sig_y)) ^ !x_mask_sign ? 1 : -1;
```
by
```
            final long __CY = bid_mult_factor[exp_diff];
            final long sig_n_prime_w1 = Math.multiplyHigh(sig_x, __CY);
            if (sig_n_prime_w1 == 0 && sig_x * __CY == sig_y)
            {
                return 0;
            }
            return (sig_n_prime_w1 == 0 && Long.compareUnsigned(sig_x * __CY, sig_y) < 0) ^ x_mask_sign ? -1 : 1;
```
and
```
        final long sig_n_prime_w1 = Mul64Impl.unsignedMultiplyHigh(sig_y, __CY);
        final long sig_n_prime_w0 = sig_y * __CY;
        // if values are equal
        if (sig_n_prime_w1 == 0 && sig_n_prime_w0 == sig_x) {
            return 0;
        }
        // if positive, return whichever significand abs is smaller
        // (converse if negative)
        return (sig_n_prime_w1 != 0 || UnsignedLong.isLess(sig_x, sig_n_prime_w0)) ^ !x_mask_sign ? 1 : -1;  
```
by
```
        final long sig_n_prime_w1 = Math.multiplyHigh(sig_y, __CY);
        if (sig_n_prime_w1 == 0 && sig_y * __CY == sig_x)
        {
            return 0;
        }
        return (sig_n_prime_w1 == 0 && Long.compareUnsigned(sig_y * __CY, sig_x) < 0) ^ x_mask_sign ? 1 : -1;
```