@plokhotnyuk 2023-07-06
Please, for yet more efficient comparison of finite numbers replace:
```
            final long sig_n_prime_w1 = Mul64Impl.unsignedMultiplyHigh(sig_x, __CY);
            final long sig_n_prime_w0 = sig_x * __CY;
            if (sig_n_prime_w1 == 0 && sig_n_prime_w0 == sig_y) {
                return 0;
            }
            return (sig_n_prime_w1 == 0 && UnsignedLong.isLess(sig_n_prime_w0, sig_y)) ^ !x_mask_sign ? 1 : -1;
```
by
```
            if (Math.multiplyHigh(sig_x, __CY) != 0) {
                return x_mask_sign ? -1 : 1;
            }
            final int cmp = Long.compareUnsigned(sig_x * __CY, sig_y);
            return x_mask_sign ? -cmp : cmp;
```
and 
```
        final long sig_n_prime_w1 = Mul64Impl.unsignedMultiplyHigh(sig_y, __CY);
        final long sig_n_prime_w0 = sig_y * __CY;
        if (sig_n_prime_w1 == 0 && sig_n_prime_w0 == sig_x) {
            return 0;
        }
        return (sig_n_prime_w1 != 0 || UnsignedLong.isLess(sig_x, sig_n_prime_w0)) ^ !x_mask_sign ? 1 : -1; 
```
by
```
        if (Math.multiplyHigh(sig_y, __CY) != 0) {
            return x_mask_sign ? 1 : -1;
        }
        final int cmp = Long.compareUnsigned(sig_y * __CY, sig_x);
        return x_mask_sign ? cmp : -cmp;
```