@agdavydov81 2023-07-27
I'm wondering: why do you invert some ternary operators? Is it just for looks or performance?